requirements.txt
Django==3.0.4
djangorestframework==3.11.0
django-cors-headers==3.2.1
pytz==2019.3
sqlparse==0.3.1
